Move SyncProductsAction and SyncVariationsAction into single class
